pyproject.toml
[project]
name = "voxtral"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "torch>=2.4.1",
  "youtube-search-python>=1.6.6",
  "tqdm>=4.66.5",
  "yt-dlp>=2024.9.27",
  "rich>=13.9.2",
  "torchaudio>=2.4.1",
  "numpy>=2.1.1",
  "huggingface-hub>=0.25.1",
  "transformers>=4.45.1",
  "einops>=0.8.0",
  "python-dotenv>=1.0.1",
  "protobuf>=5.28.2",
  "sentencepiece>=0.2.0",
  "pydantic-settings>=2.5.2",
  "wandb>=0.18.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
